requirements.txt
flask
opencv-python
mediapipe
ultralytics
yt-dlp
numpys